Moved helper function
